requirements.txt
Flask==2.2.4
jinja2==3.1.2
werkzeug==3.0.5
Flask-WTF==1.2.2
bcrypt==5.0.0
python-dotenv==1.2.2
